Fix self param into watcher notify
We must wrap the notify call in order to give access to `this` inside
watcher constructor.  This is because when `consumerMarkedDirty` is
called it is actually in the ReactiveNode context and can only pass in
the node as the self argument.  The node is not useful inside notify
because the main thing a user would want to do in here is use the
watcher object's functions like `getPending` and `watch`.

Also, there was some discussion of removing the `self` param all
together in favor of `this`.  However that does not work because in the
case of an arrow function it is defined in the global context and in the
case of a normal function we arent taking the steps to bind because it
could possibly be an arrow function.  So update the README as well to
back to the `self` way, it is better than before because `self` is actually
passed in as a param this time.

Add unit tests to describe these cases.