Fix `this` in normal function `notify`
Before `notify` was called from the `ReactiveNode` context and was
calling with the node as `this` instead of the expected watcher.
Instead, set `consumerMarkedDirty` to a function that wraps notify with
a proper call with `this`.

Also add unit tests for our expected behavior here including closure for
an arrow function.